+++ changelog.md
@@ -1,5 +1,7 @@
 ## v0.0.83
 * Add --ignore_done_flag option to Refine
+* HiveExtensions field name normalize now replaces bad SQL characters with
+  "_", not just hyphens.
 
 ## v0.0.82
 * Update hadoop, hive and spark dependency versions